requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
aiosqlite==0.20.0
curl_cffi==0.6.4
jinja2==3.1.6
openai==1.109.1
pyngrok==7.1.6
python-dotenv==1.0.1
sqlalchemy==2.0.38
telegramify-markdown==0.5.2
